feat(module-articles): introduce article management entities and module setup
Introduced a new `module-articles` package which includes entity models and module configuration necessary for article management within the application. This module includes entities such as `Article`, `ArticleCategory`, and `ArticleContent` which facilitate the organization and management of blog posts and their categorization. Additionally, the module setup integrates with `@unserve/module-user` and `@unserve/module-storage` to utilize user and storage functionalities, enhancing the module's capabilities. This setup aims to streamline the process of managing articles, improving content organization and accessibility within the application.