feat(nuxt): introduce new Nuxt module with enhanced functionality
This commit introduces a new `@unserved/nuxt` package which integrates several functionalities tailored for Nuxt applications. The package includes setup configurations for components and utilities that are automatically imported and configured. It supports Nuxt version 3.0.0 and above and includes utilities like `useClient` and `useRequest` for enhanced server communication. This module is designed to streamline the integration of Nuxt with the `@unserved` ecosystem, providing out-of-the-box support for server-client interactions and dynamic component handling.