feat(nuxt): set default `baseUrl` dynamically based on global location
Introduced dynamic determination of `baseUrl` for the global client instance in the `useClient` hook, utilizing the `useRequestURL` if available, or defaulting to `window.location.origin`. This enhancement simplifies client configuration in various runtime environments, ensuring that the `baseUrl` aligns with the current execution context, either in Nuxt or other frontend environments. This change improves flexibility and reduces configuration overhead for developers.