chore(server): add utility to check for DataSource instances
Introduced a new utility function `isDataSource` in the server package to determine if a given value is an instance of `DataSource` or meets the criteria defined for being considered as such. This utility will help in validating DataSource instances across various parts of the application, ensuring that only valid DataSource objects are processed. This update includes unit tests to verify the functionality of the utility.