chore(server): remove deprecated `createRoute` function
The `createRoute` function was marked as deprecated and has now been removed to simplify the API. This change promotes the use of `createHttpRoute` exclusively, which adheres to the current standards and best practices for defining HTTP routes in our server package. This removal is part of ongoing efforts to clean up and modernize the codebase.